requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==2.2.18
django-allauth==0.44.0
django-appconf==1.0.4
django-ckeditor==6.0.0
django-haystack==3.0
django-imagekit==4.0.2
django-js-asset==1.2.2
django-model-utils==4.1.1
django-mptt==0.12.0
django-notifications-hq==1.6.0
django-password-reset==2.0
django-taggit==1.4.0
idna==2.10
jsonfield==3.1.0
Markdown==3.3.4
numpy==1.20.2
oauthlib==3.1.0
pilkit==2.0
Pillow==8.2.0
pycparser==2.20
Pygments==2.8.1
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
swapper==1.1.2.post1
urllib3==1.26.4
